Resolve conflicts.

* origin/master: (106 commits)
  SCREAM: work around a misterious gcc8+cuda possible bug
  SCREAM: use custom gcc-8 compiler/TPL toolchain on weaver
  Changes to use larger EPS_AGRID for bi-grid ne1024pg2 (#1266)
  SCREAM: fix this pointer accesses on CUDA
  SCREAM: use test profile SHORT for cuda memcheck build
  Fix run_and_cmp tests
  Homme/SL: Add an assertion to codify 'unused'.
  Homme/SL: Fix all remaining Cuda/GCC -Wall -pedantic warnings.
  Homme/SL: Work around an nvcc issue.
  Homme(xx)/SL: Fix more -Wall -pedantic warnings.
  Homme(xx) SL: Remove a number of -Wall warnings.
  SCREAM: use test profile SHORT for valgrind builds
  SCREAM: fix uninited memory valgrind error in AtmosphereOutput
  fix alloc_prop for 2d
  remove packs from 2d remaps
  Temorarily remove ocean BGC from tested BGCEXP configurations
  Setting PERL5LIB to a shared local Perl lib for ANL machines
  Begin work on new evaluator script
  Restore SZIP_PATH and ZLIB_PATH to anlworkstation
  Remove configs for unused OSes
  ...

+++ components/homme/src/share/compose/compose_slmm_islmpi.cpp
@@ -566,7 +566,7 @@ void comm_lid_on_rank (IslMpi<MT>& cm, const Rank2Gids& rank2rmtgids,
     mpi::waitany(count, reqs.data(), &idx);
     const Int* rmtlids = recv.data() + recvptr[idx];
     const auto& gids = rank2rmtgids.at(recvrank[idx]);
-    slmm_assert(recvptr[idx+1] - recvptr[idx] == gids.size());
+    slmm_assert(recvptr[idx+1] - recvptr[idx] == static_cast<int>(gids.size()));
     Int i = 0;
     for (auto gid: gids)
       gid2rmt_owning_lid[gid] = rmtlids[i++];
